In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,13 +8,13 @@
 int main(void)
 {
 	int i, n1 = 1, n2 = 2;
-	long long int next;
+	long int next;
 
 	printf("%d, %d", n1, n2);
 	for (i = 2; i < 50; i++)
 	{
 		next = n1 + n2;
-		printf("%llu, ", next);
+		printf("%lu, ", next);
 		n1 = n2;
 		n2 = next;
 	}
